reddit_tracker/requirements.txt
praw
pandas
python-dotenv
